Properly serialize column/index tuples
This changes a few unit tests because the order of output rows changes. The
reason for that is that the output is sorted by the serialized version of the
primary key. I never made any guarantees about the order in which the rows are
returned, so this is fine for now.